backend/llm-service/requirements.txt
fastapi
uvicorn
sentence-transformers
google-generativeai
python-dotenv
pydantic
